requirements.txt
aiogram
openpyxl
requests
beautifulsoup4
aiohttp
aiohttp-cors
python-dotenv
fastapi
uvicorn